backend/requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
astunparse==1.6.3
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.5
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.2.0
ecdsa==0.19.1
fastapi==0.128.0
flatbuffers==25.12.19
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.1
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
motor==3.7.1
namex==0.1.0
narwhals==2.15.0
numpy==2.4.1
opt_einsum==3.4.0
optree==0.18.0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079150/work
pandas==2.3.3
passlib==1.7.4
patsy==1.0.2
pillow==12.1.0
protobuf==6.33.4
pyarrow==23.0.0
pyasn1==0.6.2
python-dotenv==1.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.1
six==1.17.0
smmap==5.0.2
starlette==0.50.0
statsmodels==0.14.6
streamlit==1.53.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.0.1
xgboost==3.1.3
